chore(package.json): Remove obsolete `engines` specification
For a long time we have been specifying requirements on very old
versions of Elasticsearch and NPM. Only the NPM specification was even
valid, and at this point the version numbers have fallen far behind.

We generally should support whichever NPM version comes with the
versions of Node.js we support, so it's no longer useful to specify
either of these.